Separated Go RPC code from other code:
- Identified code used by GoRPC plugins
- Created files with deprecated flags to hold code that is used by GoRPC plugins
- Added deprecation warning when loading legacy plugin
