+++ control/plugin/collector_proxy.go
@@ -19,13 +19,6 @@ limitations under the License.
 
 package plugin
 
-import (
-	"errors"
-	"fmt"
-
-	"github.com/intelsdi-x/snap/core/cdata"
-)
-
 // Arguments passed to CollectMetrics() for a Collector implementation
 type CollectMetricsArgs struct {
 	MetricTypes []MetricType
@@ -44,55 +37,4 @@ type GetMetricTypesArgs struct {
 // GetMetricTypesReply assigned by GetMetricTypes() implementation
 type GetMetricTypesReply struct {
 	MetricTypes []MetricType
-}
-
-type collectorPluginProxy struct {
-	Plugin  CollectorPlugin
-	Session Session
-}
-
-func (c *collectorPluginProxy) GetMetricTypes(args []byte, reply *[]byte) error {
-	defer catchPluginPanic(c.Session.Logger())
-
-	c.Session.Logger().Debugln("GetMetricTypes called")
-	// Reset heartbeat
-	c.Session.ResetHeartbeat()
-
-	dargs := &GetMetricTypesArgs{PluginConfig: ConfigType{ConfigDataNode: cdata.NewNode()}}
-	c.Session.Decode(args, dargs)
-
-	mts, err := c.Plugin.GetMetricTypes(dargs.PluginConfig)
-	if err != nil {
-		return errors.New(fmt.Sprintf("GetMetricTypes call error : %s", err.Error()))
-	}
-
-	r := GetMetricTypesReply{MetricTypes: mts}
-	*reply, err = c.Session.Encode(r)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *collectorPluginProxy) CollectMetrics(args []byte, reply *[]byte) error {
-	defer catchPluginPanic(c.Session.Logger())
-	c.Session.Logger().Debugln("CollectMetrics called")
-	// Reset heartbeat
-	c.Session.ResetHeartbeat()
-
-	dargs := &CollectMetricsArgs{}
-	c.Session.Decode(args, dargs)
-
-	ms, err := c.Plugin.CollectMetrics(dargs.MetricTypes)
-	if err != nil {
-		return errors.New(fmt.Sprintf("CollectMetrics call error : %s", err.Error()))
-	}
-
-	r := CollectMetricsReply{PluginMetrics: ms}
-	*reply, err = c.Session.Encode(r)
-	if err != nil {
-		return err
-	}
-	return nil
 }
