+++ control/plugin/metric.go
@@ -23,11 +23,7 @@ import (
 	"bytes"
 	"encoding/gob"
 	"encoding/json"
-	"errors"
-	"fmt"
 	"time"
-
-	log "github.com/Sirupsen/logrus"
 
 	"github.com/intelsdi-x/snap/core"
 	"github.com/intelsdi-x/snap/core/cdata"
@@ -42,8 +38,6 @@ const (
 	SnapGOBContentType = "snap.gob"
 	// SnapJSON snap metrics serialized into json
 	SnapJSONContentType = "snap.json"
-	// SnapProtoBuff snap metrics serialized into protocol buffers
-	// SnapProtoBuff = "snap.pb" // TO BE IMPLEMENTED
 )
 
 type ConfigType struct {
@@ -81,12 +75,6 @@ func (p *ConfigType) GobDecode(data []byte) error {
 	return nil
 }
 
-func NewPluginConfigType() ConfigType {
-	return ConfigType{
-		ConfigDataNode: cdata.NewNode(),
-	}
-}
-
 // Represents a metric type. Only used within plugins and across plugin calls.
 // Converted to core.MetricType before being used within modules.
 type MetricType struct {
@@ -120,18 +108,6 @@ type MetricType struct {
 
 	// The timestamp from when the metric was created.
 	Timestamp_ time.Time `json:"timestamp"`
-}
-
-// NewMetricType returns a Constructor
-func NewMetricType(namespace core.Namespace, timestamp time.Time, tags map[string]string, unit string, data interface{}) *MetricType {
-	return &MetricType{
-		Namespace_:          namespace,
-		Tags_:               tags,
-		Data_:               data,
-		Timestamp_:          timestamp,
-		LastAdvertisedTime_: timestamp,
-		Unit_:               unit,
-	}
 }
 
 // Returns the namespace.
@@ -181,120 +157,4 @@ func (p MetricType) Unit() string {
 
 func (p *MetricType) AddData(data interface{}) {
 	p.Data_ = data
-}
-
-// MarshalMetricTypes returns a []byte containing a serialized version of []MetricType using the content type provided.
-func MarshalMetricTypes(contentType string, metrics []MetricType) ([]byte, string, error) {
-	// If we have an empty slice we return an error
-	if len(metrics) == 0 {
-		es := fmt.Sprintf("attempt to marshall empty slice of metrics: %s", contentType)
-		log.WithFields(log.Fields{
-			"_module": "control-plugin",
-			"block":   "marshal-content-type",
-			"error":   es,
-		}).Error("error while marshalling")
-		return nil, "", errors.New(es)
-	}
-	// Switch on content type
-	switch contentType {
-	case SnapAllContentType, SnapGOBContentType:
-		// NOTE: A snap All wildcard will result in GOB
-		var buf bytes.Buffer
-		enc := gob.NewEncoder(&buf)
-		err := enc.Encode(metrics)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"_module": "control-plugin",
-				"block":   "marshal-content-type",
-				"error":   err.Error(),
-			}).Error("error while marshalling")
-			return nil, "", err
-		}
-		// contentType := SnapGOBContentType
-		return buf.Bytes(), SnapGOBContentType, nil
-	case SnapJSONContentType:
-		// Serialize into JSON
-		b, err := json.Marshal(metrics)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"_module": "control-plugin",
-				"block":   "marshal-content-type",
-				"error":   err.Error(),
-			}).Error("error while marshalling")
-			return nil, "", err
-		}
-		return b, SnapJSONContentType, nil
-	default:
-		// We don't recognize this content type. Log and return error.
-		es := fmt.Sprintf("invalid snap content type: %s", contentType)
-		log.WithFields(log.Fields{
-			"_module": "control-plugin",
-			"block":   "marshal-content-type",
-			"error":   es,
-		}).Error("error while marshalling")
-		return nil, "", errors.New(es)
-	}
-}
-
-// UnmarshallMetricTypes takes a content type and []byte payload and returns a []MetricType
-func UnmarshallMetricTypes(contentType string, payload []byte) ([]MetricType, error) {
-	switch contentType {
-	case SnapGOBContentType:
-		var metrics []MetricType
-		r := bytes.NewBuffer(payload)
-		err := gob.NewDecoder(r).Decode(&metrics)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"_module": "control-plugin",
-				"block":   "unmarshal-content-type",
-				"error":   err.Error(),
-			}).Error("error while unmarshalling")
-			return nil, err
-		}
-		return metrics, nil
-	case SnapJSONContentType:
-		var metrics []MetricType
-		err := json.Unmarshal(payload, &metrics)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"_module": "control-plugin",
-				"block":   "unmarshal-content-type",
-				"error":   err.Error(),
-			}).Error("error while unmarshalling")
-			return nil, err
-		}
-		return metrics, nil
-	default:
-		// We don't recognize this content type as one we can unmarshal. Log and return error.
-		es := fmt.Sprintf("invalid snap content type for unmarshalling: %s", contentType)
-		log.WithFields(log.Fields{
-			"_module": "control-plugin",
-			"block":   "unmarshal-content-type",
-			"error":   es,
-		}).Error("error while unmarshalling")
-		return nil, errors.New(es)
-	}
-}
-
-// SwapMetricContentType swaps a payload with one content type to another one.
-func SwapMetricContentType(contentType, requestedContentType string, payload []byte) ([]byte, string, error) {
-	metrics, err1 := UnmarshallMetricTypes(contentType, payload)
-	if err1 != nil {
-		log.WithFields(log.Fields{
-			"_module": "control-plugin",
-			"block":   "swap-content-type",
-			"error":   err1.Error(),
-		}).Error("error while swaping")
-		return nil, "", err1
-	}
-	newPayload, newContentType, err2 := MarshalMetricTypes(requestedContentType, metrics)
-	if err2 != nil {
-		log.WithFields(log.Fields{
-			"_module": "control-plugin",
-			"block":   "swap-content-type",
-			"error":   err2.Error(),
-		}).Error("error while swaping")
-		return nil, "", err2
-	}
-	return newPayload, newContentType, nil
 }
