+++ control/plugin/publisher_proxy.go
@@ -19,56 +19,10 @@ limitations under the License.
 
 package plugin
 
-import (
-	"errors"
-	"fmt"
-
-	"github.com/intelsdi-x/snap/core/ctypes"
-)
+import "github.com/intelsdi-x/snap/core/ctypes"
 
 type PublishArgs struct {
 	ContentType string
 	Content     []byte
 	Config      map[string]ctypes.ConfigValue
 }
-
-type PublishReply struct {
-}
-
-type publisherPluginProxy struct {
-	Plugin  PublisherPlugin
-	Session Session
-}
-
-func (p *publisherPluginProxy) Publish(args []byte, reply *[]byte) error {
-	defer catchPluginPanic(p.Session.Logger())
-	p.Session.ResetHeartbeat()
-
-	dargs := &PublishArgs{}
-	err := p.Session.Decode(args, dargs)
-	if err != nil {
-		return err
-	}
-
-	err = p.Plugin.Publish(dargs.ContentType, dargs.Content, dargs.Config)
-	if err != nil {
-		return errors.New(fmt.Sprintf("Publish call error: %v", err.Error()))
-	}
-	return nil
-}
-
-/*
-type gRPCPublisherProxy struct {
-	Plugin  PublisherPlugin
-	Session Session
-	gRPCPluginProxy
-}
-
-func (p *gRPCPublisherProxy) Publish(ctx context.Context, arg *rpc.PublishArg) (*common.Empty, error) {
-	defer catchPluginPanic(p.Session.Logger())
-	err := p.Plugin.Publish(arg.ContentType, arg.Content, common.ParseConfig(arg.Config))
-	if err != nil {
-		return &common.Empty{}, err
-	}
-	return &common.Empty{}, nil
-}*/
